minor tweaks and fixture updates
Addressed the final customization issues we wanted to implement - chiefly, changing the language in the Search bar and adding an "Annotation" fixture from a custom metadata field. Also updated the suite of local text fixtures with the added "annotation" field.

The only remaining things to do at this point are to deal with the remaining Dependabot alerts.