updated fixtures to align with aardvark
A few tweaks to the aardvark schema have been made. Updated the UMass fixtures to align with them.

Note that I attempted to update the fixtures for the GBL documents (which were updated in the summer 2021 sprint) but they broke my version of GBL. I assume I'll need to upgrade to 3.4 at the least to get them to work properly.